Fixed some responsive issues
Created the SizeWatcher object which will reactively watch the size of an element that it is given. It is used to contain the current size of an element at any point in time (without spamming event triggers).

Updated the bars in the title background to use vw units (much cleaner - more performant).

Removed the now unused VueProject component.