Created a new & better background title animation
The text in the background now changes 1 character at a time between different lines of code

getBarStartY() and getBarHeight() should be treated as constants. As they were, they were reactively updating (which is causing issues since their values are based on randomization). Now they are assigned a value only on mount.