Updated CodeLine to be more performant
CodeLine now updates 3 characters at a time instead of 1, but the timer runs at 500 instead of 200, so the methods are called less often.

Updated the CodeLines in TitleBackground.vue to be a reactive() object to fix reactivity issues (the reactivity fix also involves updating the watcher to watch the object).

Split the functionality in CodeLine.ts into multiple functions (since the code bloated a little bit more in order to allow it to handle multiple characters at once).