Better animated the project carousel
I had to do a "hack" of having a duplicate image element and duplicate text (animating one with gsap - the other stays static) in order to achieve the animations I wanted (since GSAP was reversing the opacity of all of the animations to 0 for some reason).

I also removed a paragraph of SandPaper's description (since the description was taking up too much space).