Added gradients to the title bars
This ended up being more work than I expected.

Created some classes to help encapsulate the data

Separated the title bars into their own foreground component

The effect is currently not animated. It is broken with a reactivity issue. Unfortunately, I am very short on time to finish this website, so I am leaving that effect disabled indefinitely (since it is not that important).